Update _config.yml
Amended line 17 of the description for a more SEO optimised paragraph. 

description: 'Welcome to the homepage of the Torr Vision Group at the University of Oxford'

to

description: 'Welcome to the Torr Vision Group (TVG) at the University of Oxford Department of Engineering Science. We are led by Professor Philip Torr and engage in state-of-the-art research into the mathematical theory of computer vision and artificial intelligence, relevant to the needs of society. View our published research papers, postdoctoral research opportunities and recruitment, prizes, and the members of our team.'